Migrate machine test
